+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+[Unreleased]
+###
+- Added `parameters.pkg.preferGlobal` parsing.
+
 [0.4.0] - 2017-12-10
 --------------------
 ### Added
@@ -31,6 +35,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial release.
 
+[Unreleased]: https://github.com/jbenner-radham/node-readme-md/compare/v0.4.0...HEAD
 [0.4.0]: https://github.com/jbenner-radham/node-readme-md/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/jbenner-radham/node-readme-md/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/jbenner-radham/node-readme-md/compare/v0.1.0...v0.2.0